# src/main/java/org/kiss/model/entity/util/CarLocation.java
package org.kiss.model.entity.util;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CarLocation {

    private String city;
    private String country;
}