# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Car Rental API
  description: Backend API for the car rental application.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/cars:
    get:
      tags:
        - Cars
      operationId: listCars
      summary: Search available car listings. Public, paginated.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: companyId
          in: query
          description: Return only cars belonging to this company.
          schema:
            type: string
        - name: make
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/CarCategory"
        - name: transmission
          in: query
          schema:
            $ref: "#/components/schemas/TransmissionType"
        - name: fuelType
          in: query
          schema:
            $ref: "#/components/schemas/FuelType"
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: city
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: availableFrom
          in: query
          description: >
            Only return cars with no reservation conflicting with [availableFrom, availableUntil)
            (a 2-hour buffer is applied around existing reservations, same rule as reservation
            creation). Must be provided together with availableUntil.
          schema:
            type: string
            format: date-time
        - name: availableUntil
          in: query
          description: See availableFrom. Must be after availableFrom.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: A page of car listings matching the search criteria.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarPage"
        "400":
          description: >
            Invalid pagination or filter parameters, or availableFrom/availableUntil were not
            provided together, or availableUntil is not after availableFrom.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Cars
      operationId: registerCar
      summary: Register a new car listing. Restricted to dealership accounts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarRegistrationRequest"
      responses:
        "201":
          description: Car registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not a dealership account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A car with this VIN is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cars/mine:
    get:
      tags:
        - Cars
      operationId: listMyCars
      summary: List the cars registered by the currently authenticated dealership.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of the dealership's car listings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarPage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not a dealership account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cars/{id}:
    get:
      tags:
        - Cars
      operationId: getCarById
      summary: Get the details of a single car listing. Public.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The car listing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponse"
        "404":
          description: No car exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Cars
      operationId: updateCar
      summary: Update a car listing. Admins may update any car; dealership accounts may only update a car belonging to their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdateRequest"
      responses:
        "200":
          description: Car updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin, or a dealership account whose company does not own this car.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No car exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Cars
      operationId: deleteCar
      summary: >
        Delete a car. Admins may delete any car; dealership accounts may only delete a car
        belonging to their own company.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Car deleted.
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin, or a dealership account whose company does not own this car.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No car exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/car-types:
    get:
      tags:
        - CarTypes
      operationId: listCarTypes
      summary: List all car types. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of car types.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarTypePage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - CarTypes
      operationId: createCarType
      summary: >
        Create a car type. Restricted to admin accounts. The response includes a presigned URL
        the frontend can use to upload the preview image directly to storage.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarTypeRequest"
      responses:
        "201":
          description: Car type created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarTypeCreationResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A car type with this make and model already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/car-types/{id}:
    delete:
      tags:
        - CarTypes
      operationId: deleteCarType
      summary: Delete a car type. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Car type deleted.
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No car type exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - timestamp
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    CarCategory:
      type: string
      enum:
        - SEDAN
        - SUV
        - HATCHBACK
        - COUPE
        - CONVERTIBLE
        - MINIVAN
        - PICKUP
        - VAN
    TransmissionType:
      type: string
      enum:
        - MANUAL
        - AUTOMATIC
    FuelType:
      type: string
      enum:
        - PETROL
        - DIESEL
        - ELECTRIC
        - HYBRID
        - LPG
    CarStatus:
      type: string
      description: Listing status, controlled by the dealership.
      enum:
        - AVAILABLE
        - RENTED
        - MAINTENANCE
        - UNLISTED
    CarLocation:
      type: object
      required:
        - city
        - country
      properties:
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
    CarRegistrationRequest:
      type: object
      required:
        - make
        - model
        - year
        - category
        - transmission
        - fuelType
        - seats
        - mileageKm
        - dailyPriceAmount
        - location
      properties:
        companyId:
          type: string
          description: >
            Company to register this car for. Required when an admin registers the car. Ignored when a
            dealership registers their own car — the car is instead attached to the dealership's own company.
        vin:
          type: string
          maxLength: 17
        licensePlate:
          type: string
          maxLength: 20
        make:
          type: string
          minLength: 1
          description: Must match an existing CarType together with model.
        model:
          type: string
          minLength: 1
          description: Must match an existing CarType together with make.
        year:
          type: integer
          minimum: 1900
          maximum: 2100
        category:
          $ref: "#/components/schemas/CarCategory"
        transmission:
          $ref: "#/components/schemas/TransmissionType"
        fuelType:
          $ref: "#/components/schemas/FuelType"
        seats:
          type: integer
          minimum: 1
          maximum: 20
        doors:
          type: integer
          minimum: 1
          maximum: 6
        color:
          type: string
        mileageKm:
          type: integer
          minimum: 0
        dailyPriceAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        dailyPriceCurrency:
          type: string
          default: EUR
        location:
          $ref: "#/components/schemas/CarLocation"
        features:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
        description:
          type: string
          maxLength: 2000
    CarUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/CarRegistrationRequest"
        - type: object
          properties:
            status:
              $ref: "#/components/schemas/CarStatus"
    CarResponse:
      type: object
      required:
        - id
        - companyId
        - make
        - model
        - year
        - category
        - transmission
        - fuelType
        - seats
        - mileageKm
        - dailyPriceAmount
        - dailyPriceCurrency
        - status
        - location
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        companyId:
          type: string
          description: >
            The company this car belongs to — set by the dealership's own company when a dealership
            registers the car, or by the admin-supplied companyId when an admin registers it.
        vin:
          type: string
        licensePlate:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        category:
          $ref: "#/components/schemas/CarCategory"
        transmission:
          $ref: "#/components/schemas/TransmissionType"
        fuelType:
          $ref: "#/components/schemas/FuelType"
        seats:
          type: integer
        doors:
          type: integer
        color:
          type: string
        mileageKm:
          type: integer
        dailyPriceAmount:
          type: number
          format: double
        dailyPriceCurrency:
          type: string
        status:
          $ref: "#/components/schemas/CarStatus"
        location:
          $ref: "#/components/schemas/CarLocation"
        features:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CarPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CarResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    CarTypeRequest:
      type: object
      required:
        - make
        - model
      properties:
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
    CarTypeResponse:
      type: object
      required:
        - id
        - make
        - model
        - previewUrl
      properties:
        id:
          type: string
        make:
          type: string
        model:
          type: string
        previewUrl:
          type: string
          description: >
            Public URL where the preview image will be served from once uploaded. Set immediately
            at creation time, before the image is actually uploaded.
    CarTypeCreationResponse:
      type: object
      required:
        - carType
        - previewUploadUrl
      properties:
        carType:
          $ref: "#/components/schemas/CarTypeResponse"
        previewUploadUrl:
          type: string
          description: >
            Presigned URL the frontend must PUT the preview image bytes to directly (bypassing
            this API). Valid for 10 minutes, single use.
    CarTypePage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CarTypeResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...